Fix some unlogic moments 1

+++ averich.vladimir/I2/istriplet.cpp
@@ -1,16 +1,17 @@
 #include "istriplet.hpp"
 #include <iostream>
-int isTriplet(int arr[], int size)
+#include <cstddef>
+int isTriplet(int arr[], size_t size)
 {
   if (size < 0)
   {
    std::cout << "Error: size of array is wrong, try again";
   }
-  int i = 0;
-  int j = 1;
-  int k = 2;
+  size_t i = 0;
+  size_t j = 1;
+  size_t k = 2;
   int countOfTriplet = 0;
-  for (int l = 0; l < size; l++)
+  for (size_t l = 0; l < size; l++)
   {
     if (arr[i] > arr[j] && arr[j] > arr[k])
     {
